restore original LocalBGS and leave LocalBGSProxy to overwrite it in the future
