+++ modules/claims_api/spec/requests/v2/veterans/claims_request_spec.rb
@@ -42,7 +42,6 @@ RSpec.describe 'Claims', type: :request do
     before do
       allow(Flipper).to receive(:enabled?).with(:claims_status_v2_lh_benefits_docs_service_enabled).and_return false
       allow(Flipper).to receive(:enabled?).with(:claims_load_testing).and_return false
-      allow(Flipper).to receive(:enabled?).with(:claims_api_local_bgs_refactor).and_return false
     end
 
     describe 'index' do
